Add files via upload
The Income dataset contains a number of variables from the US census. The data file provided has a header describing the fields in the file, and approximately 32000 rows of data. The aim is to create a classifier using Spark + MLLib that can determine whether a person earns above $50k given the other variables in the data set. Due to the nature of certain fields, any categorical fields have to be converted into indexes for an improved performance of the machine learning algorithms.
The Random Forest and Decision Tree classifiers are used, and the aim is to get a fairly high accuracy (>75%)
A confusion matrix is added to take account of the limitation of the accuracy and provide model performance analysis and improve predictive accuracy.