Add files via upload
This project is a clone development for the popular Wordle game. 
The dataset provided contains a list of all 5-letter words for my wordle clone to randomly pull the secret word from - should be the same as the actual wordle game list.
The algorithm is set to inform you what the secret word is before playing (this was just for testing), but please do comment that line of the code and play for a bit more of a challenge. Happy Guessing!!!
NOTE: Present and correctly positioned letters are displayed as 'uppercase letters' and in the colour 'green'; present and incorrectly positioned letters are displayed as 'lowercase letters' and in the colour 'yellow'; And letters that are not present at all, are displayed as a 'full stop' and in the colour 'grey'.